refactor: moved str_before function
